requirements.txt
asgiref==3.5.2
astroid==2.11.7
async-generator==1.10
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
colorama==0.4.5
cryptography==36.0.1
dill==0.3.5.1
distlib==0.3.5
Django==4.1
et-xmlfile==1.1.0
filelock==3.8.0
h11==0.13.0
idna==3.3
isort==5.10.1
jarowinkler==1.0.2
lazy-object-proxy==1.7.1
lxml==4.6.4
mccabe==0.7.0
mysqlclient==2.1.1
numpy==1.23.3
opencv-python==4.6.0.66
openpyxl==3.0.9
outcome==1.1.0
pandas==1.4.1
Pillow==9.2.0
pipenv==2022.4.21
platformdirs==2.5.2
pptx2md==1.1.2
pycparser==2.21
pylint==2.14.5
pyOpenSSL==22.0.0
PySimpleGUI==4.60.3
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.21.0
python-pptx==0.6.21
pytz==2022.1
rapidfuzz==2.0.7
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
tomli==2.0.1
tomlkit==0.11.4
trio==0.19.0
trio-websocket==0.9.2
tzdata==2022.2
urllib3==1.26.8
virtualenv==20.16.3
wrapt==1.14.1
wsproto==1.0.0
XlsxWriter==3.0.3
